libs/AI Agent In Practice/05_Multi-Agent 系统/5.4 _动手_搭建代码生成+审查的双 Agent 系统/requirements.txt
litellm>=1.40.0
python-dotenv>=1.0.0
pytest>=7.0.0
pylint>=3.0.0
flake8>=7.0.0
bandit>=1.7.0
